refactor: extracted activity switcher & moved onclick event
Refs: #UPS-5078
