+++ src/shared/lib/utils/hooks/index.ts
@@ -3,3 +3,4 @@ export * from "./useIsBlacklisted";
 export * from "./useMenu";
 export * from "./useIsPreviousRender";
 export * from "./useDetectMobile";
+export * from "./useTranslation";
